move the RocDecVcnCodecSpec class to a new header file
